One branch to rule them all!
This is simple. What we're doing here is grep-ing the manifest for
o8x-caf (our caf branch) and telling it to use X roomservice and if not
use the other one.

This keeps us from having to use two branches in the build repo, especially
since we're now using the vendor/extras.

It could be prettier for sure (I'm not a script wiz) but this works :-)

*Tested with bullhead and oneplus3*

Change-Id: I4bea9bbc087e26273e34a216d05d879ef665bcaa
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>